+++ nodes/blind-control.js
@@ -8,6 +8,7 @@ const ctrlLib = require(path.join(__dirname, '/lib/timeControlHelper.js'));
 const util = require('util');
 const clonedeep = require('lodash.clonedeep');
 const isEqual = require('lodash.isequal');
+const { readFile } = require('fs');
 
 const cRuleUntil = 0;
 const cRuleFrom = 1;
@@ -271,7 +272,7 @@ module.exports = function (RED) {
                 node.nodeData.overwrite.importance = nImportance;
             }
             node.nodeData.overwrite.active = true;
-            node.context().set('overwrite', node.nodeData.overwrite, node.storeName);
+            node.context().set('overwrite', node.nodeData.overwrite, node.contextStore);
         } else if (node.nodeData.overwrite.active) {
             node.debug(`overwrite active, check of nImportance=${nImportance} or nExpire=${nExpire}`);
             if (Number.isFinite(nExpire)) {
@@ -283,7 +284,7 @@ module.exports = function (RED) {
                 // set to new importance
                 node.nodeData.overwrite.importance = nImportance;
             }
-            node.context().set('overwrite', node.nodeData.overwrite, node.storeName);
+            node.context().set('overwrite', node.nodeData.overwrite, node.contextStore);
         }
         // node.debug(`overwrite exit node.nodeData.overwrite.active=${node.nodeData.overwrite.active}; expire=${nExpire};  newPos=${newPos}`);
         return ctrlLib.setOverwriteReason(node);
@@ -484,6 +485,24 @@ module.exports = function (RED) {
     }
     /******************************************************************************************/
     /**
+     * changes the rule settings
+     * @param {Object} node node data
+     * @param {number} [rulePos] the position of the rule which should be changed
+     * @param {string} [ruleName] the name of the rule which should be changed
+     * @param {Object} ruleData the properties of the rule which should be changed
+     */
+    function changeRules(node, rulePos, ruleName, ruleData) {
+        node.debug(`changeRules: ${ node.rules.count } ruleData:' ${util.inspect(ruleData, {colors:true, compact:10})}`);
+        for (let i = 0; i <= node.rules.count; ++i) {
+            const rule = node.rules[i];
+            if (((typeof rulePos !== 'undefined') && rule.pos === rulePos) ||
+                ((typeof ruleName !== 'undefined') && rule.name === ruleName)) {
+                node.rules[i] = Object.assign(node.rules[i], ruleData);
+            }
+        }
+    }
+    /******************************************************************************************/
+    /**
      * check all rules and determinate the active rule
      * @param {Object} node node data
      * @param {Object} msg the message object
@@ -507,6 +526,7 @@ module.exports = function (RED) {
         for (let i = 0; i <= node.rules.lastUntil; ++i) {
             const rule = node.rules.data[i];
             // node.debug('rule ' + rule.time.operator + ' - ' + (rule.time.operator !== cRuleFrom) + ' - ' + util.inspect(rule, {colors:true, compact:10, breakLength: Infinity }));
+            if (!rule.enabled) { continue; }
             if (rule.time && rule.time.operator === cRuleFrom) { continue; }
             // const res = fktCheck(rule, r => (r >= nowNr));
             let res = null;
@@ -539,6 +559,7 @@ module.exports = function (RED) {
             // node.debug('--------- starting second loop ' + node.rules.count);
             for (let i = (node.rules.count - 1); i >= 0; --i) {
                 const rule = node.rules.data[i];
+                if (!rule.enabled) { continue; }
                 // node.debug('rule ' + rule.time.operator + ' - ' + (rule.time.operator !== cRuleUntil) + ' - ' + util.inspect(rule, {colors:true, compact:10, breakLength: Infinity }));
                 if (rule.time && rule.time.operator === cRuleUntil) { continue; } // - From: timeOp === cRuleFrom
                 const res = ctrlLib.compareRules(node, msg, rule, r => (r <= oNow.nowNr), oNow);
@@ -790,7 +811,7 @@ module.exports = function (RED) {
         };
         // temporary node Data
         node.levelReverse = false;
-        node.storeName = config.storeName || '';
+        node.contextStore = config.contextStore || this.positionConfig.contextStore;
         // Retrieve the config node
         node.sunData = {
             /** Defines if the sun control is active or not */
@@ -840,7 +861,7 @@ module.exports = function (RED) {
             addId: config.addId,
             addIdType: config.addIdType||'none',
             /** The override settings */
-            overwrite: node.context().get('overwrite', node.storeName) || {
+            overwrite: node.context().get('overwrite', node.contextStore) || {
                 active: false,
                 importance: 0
             }
@@ -982,7 +1003,7 @@ module.exports = function (RED) {
             current: NaN, // unknown
             currentInverse: NaN
         };
-        node.previousData = node.context().get('previous', node.storeName) || {
+        node.previousData = node.context().get('previous', node.contextStore) || {
             level: NaN, // unknown
             reasonCode: -1,
             usedRule: NaN,
@@ -1055,7 +1076,7 @@ module.exports = function (RED) {
                 if (Number.isFinite(newMode) && newMode >= 0 && newMode <= node.sunData.modeMax) {
                     node.debug(`set mode from ${node.sunData.mode} to ${newMode}`);
                     node.sunData.mode = newMode;
-                    node.context().set('mode', newMode, node.storeName);
+                    node.context().set('mode', newMode, node.contextStore);
                 }
 
                 if (msg.topic && (typeof msg.topic === 'string') && msg.topic.startsWith('set')) {
@@ -1092,8 +1113,20 @@ module.exports = function (RED) {
                         case 'setAutoTriggerTime':
                             node.autoTrigger.defaultTime = parseInt(msg.payload) || node.autoTrigger.defaultTime;
                             break;
-                        case 'setStoreName':
-                            node.storeName = msg.payload || node.storeName;
+                        case 'setContextStore':
+                            node.contextStore = msg.payload || node.contextStore;
+                            break;
+                        case 'disableRule':
+                            changeRules(node, undefined, msg.payload, { enabled: false });
+                            break;
+                        case 'enableRule':
+                            changeRules(node, undefined, msg.payload, { enabled: true });
+                            break;
+                        case 'disableRuleByPos':
+                            changeRules(node, parseInt(msg.payload), undefined, { enabled: false });
+                            break;
+                        case 'enableRuleByPos':
+                            changeRules(node, parseInt(msg.payload), undefined, { enabled: true });
                             break;
                         default:
                             break;
@@ -1108,7 +1141,7 @@ module.exports = function (RED) {
 
                 // initialize
                 node.nowarn = {};
-                const tempData = node.context().get('cacheData',node.storeName) || {};
+                const tempData = node.context().get('cacheData',node.contextStore) || {};
                 if (!isNaN(node.level.current)) {
                     node.previousData.level = node.level.current;
                     node.previousData.levelInverse = node.level.currentInverse;
@@ -1117,7 +1150,7 @@ module.exports = function (RED) {
                     node.previousData.reasonCode = node.reason.code;
                     node.previousData.reasonState = node.reason.state;
                     node.previousData.reasonDescription = node.reason.description;
-                    node.context().set('previous', node.previousData, node.storeName);
+                    node.context().set('previous', node.previousData, node.contextStore);
                 }
                 node.oversteer.isChecked = false;
                 node.reason.code = NaN;
@@ -1157,7 +1190,7 @@ module.exports = function (RED) {
                     node.previousData.last.ruleId = blindCtrl.rule.id;
                     node.previousData.last.ruleLevel = blindCtrl.rule.level;
                     node.previousData.last.ruleTopic = blindCtrl.rule.topic;
-                    node.context().set('previous', node.previousData, node.storeName);
+                    node.context().set('previous', node.previousData, node.contextStore);
                     if (blindCtrl.rule.isOff === true) {
                         // rule set the controller off
                         done();
@@ -1264,7 +1297,7 @@ module.exports = function (RED) {
                     state: node.reason.state,
                     description: node.reason.description,
                     rule: ruleId,
-                    mode: node.context().get('mode', node.storeName) || node.sunData.mode,
+                    mode: node.context().get('mode', node.contextStore) || node.sunData.mode,
                     topic: msg.topic,
                     payload: msg.payload
                 };
@@ -1309,7 +1342,7 @@ module.exports = function (RED) {
                 if (isNaN(ruleId)) {
                     node.previousData.usedRule = ruleId;
                 }
-                node.context().set('cacheData', tempData, node.storeName);
+                node.context().set('cacheData', tempData, node.contextStore);
                 if (node.autoTrigger) {
                     node.debug('next autoTrigger will set to ' + node.autoTrigger.time + ' - ' + node.autoTrigger.type);
                     if (node.autoTriggerObj) {
@@ -1362,4 +1395,24 @@ module.exports = function (RED) {
     }
 
     RED.nodes.registerType('blind-control', sunBlindControlNode);
+
+    RED.httpAdmin.post('/blind-control/:id', RED.auth.needsPermission('blind-control.write'), (req, res) => {
+        const node = RED.nodes.getNode(req.params.id);
+        if (node !== null && typeof node !== 'undefined') {
+            try {
+                if (req.body && req.body.ruleData) {
+                    changeRules(node, req.body.rulePos, req.body.ruleName, req.body.ruleData);
+                }
+                res.sendStatus(200);
+            } catch(err) {
+                res.sendStatus(500);
+                node.error(RED._('node-red:inject.failed',{error:err.toString()}));
+            }
+        } else {
+            res.status(404).send(JSON.stringify({
+                error: 'can not find node "' +req.params.id+'" '+String(node)
+            }));
+            return;
+        }
+    });
 };