+++ nodes/clock-timer.js
@@ -106,7 +106,7 @@ module.exports = function (RED) {
                 node.nodeData.overwrite.importance = nImportance;
             }
             node.nodeData.overwrite.active = true;
-            node.context().set('overwrite', node.nodeData.overwrite, node.storeName);
+            node.context().set('overwrite', node.nodeData.overwrite, node.contextStore);
         } else if (node.nodeData.overwrite.active) {
             node.debug(`overwrite active, check of nImportance=${nImportance} or nExpire=${nExpire}`);
             if (Number.isFinite(nExpire)) {
@@ -118,12 +118,30 @@ module.exports = function (RED) {
                 // set to new importance
                 node.nodeData.overwrite.importance = nImportance;
             }
-            node.context().set('overwrite', node.nodeData.overwrite, node.storeName);
+            node.context().set('overwrite', node.nodeData.overwrite, node.contextStore);
         }
         // node.debug(`overwrite exit node.nodeData.overwrite.active=${node.nodeData.overwrite.active}; expire=${nExpire}`);
         return ctrlLib.setOverwriteReason(node);
     }
 
+    /******************************************************************************************/
+    /**
+     * changes the rule settings
+     * @param {Object} node node data
+     * @param {number} [rulePos] the position of the rule which should be changed
+     * @param {string} [ruleName] the name of the rule which should be changed
+     * @param {Object} ruleData the properties of the rule which should be changed
+     */
+    function changeRules(node, rulePos, ruleName, ruleData) {
+        node.debug(`changeRules: ${ node.rules.count } ruleData:' ${util.inspect(ruleData, {colors:true, compact:10})}`);
+        for (let i = 0; i <= node.rules.count; ++i) {
+            const rule = node.rules[i];
+            if (((typeof rulePos !== 'undefined') && rule.pos === rulePos) ||
+                ((typeof ruleName !== 'undefined') && rule.name === ruleName)) {
+                node.rules[i] = Object.assign(node.rules[i], ruleData);
+            }
+        }
+    }
     /******************************************************************************************/
     /**
      * check all rules and determinate the active rule
@@ -144,6 +162,7 @@ module.exports = function (RED) {
         // node.debug('first loop count:' + node.rules.count + ' lastuntil:' + node.rules.lastUntil);
         for (let i = 0; i <= node.rules.lastUntil; ++i) {
             const rule = node.rules.data[i];
+            if (!rule.enabled) { continue; }
             // node.debug('rule ' + rule.time.operator + ' - ' + (rule.time.operator !== cRuleFrom) + ' - ' + util.inspect(rule, {colors:true, compact:10, breakLength: Infinity }));
             if (rule.time && rule.time.operator === cRuleFrom) { continue; }
             // const res = fktCheck(rule, r => (r >= nowNr));
@@ -167,6 +186,7 @@ module.exports = function (RED) {
             // node.debug('--------- starting second loop ' + node.rules.count);
             for (let i = (node.rules.count - 1); i >= 0; --i) {
                 const rule = node.rules.data[i];
+                if (!rule.enabled) { continue; }
                 // node.debug('rule ' + rule.time.operator + ' - ' + (rule.time.operator !== cRuleUntil) + ' - ' + util.inspect(rule, {colors:true, compact:10, breakLength: Infinity }));
                 if (rule.time && rule.time.operator === cRuleUntil) { continue; } // - From: timeOp === cRuleFrom
                 const res = ctrlLib.compareRules(node, msg, rule, r => (r <= oNow.nowNr), oNow);
@@ -327,7 +347,7 @@ module.exports = function (RED) {
             description: ''
         };
         // temporary node Data
-        node.storeName = config.storeName || '';
+        node.contextStore = config.contextStore || this.positionConfig.contextStore;
         node.nodeData = {
             /** The Level of the window */
             payloadDefault: config.payloadDefault,
@@ -340,7 +360,7 @@ module.exports = function (RED) {
             addId: config.addId,
             addIdType: config.addIdType||'none',
             /** The override settings */
-            overwrite: node.context().get('overwrite', node.storeName) || {
+            overwrite: node.context().get('overwrite', node.contextStore) || {
                 active: false,
                 importance: 0
             }
@@ -423,8 +443,20 @@ module.exports = function (RED) {
                         case 'setAutoTriggerTime':
                             node.autoTrigger.defaultTime = parseInt(msg.payload) || node.autoTrigger.defaultTime;
                             break;
-                        case 'setStoreName':
-                            node.storeName = msg.payload || node.storeName;
+                        case 'setCntextStore':
+                            node.contextStore = msg.payload || node.contextStore;
+                            break;
+                        case 'disableRule':
+                            changeRules(node, undefined, msg.payload, { enabled: false });
+                            break;
+                        case 'enableRule':
+                            changeRules(node, undefined, msg.payload, { enabled: true });
+                            break;
+                        case 'disableRuleByPos':
+                            changeRules(node, parseInt(msg.payload), undefined, { enabled: false });
+                            break;
+                        case 'enableRuleByPos':
+                            changeRules(node, parseInt(msg.payload), undefined, { enabled: true });
                             break;
                         default:
                             break;
@@ -433,7 +465,7 @@ module.exports = function (RED) {
 
                 // initialize
                 node.nowarn = {};
-                const tempData = node.context().get('cacheData',node.storeName) || {};
+                const tempData = node.context().get('cacheData',node.contextStore) || {};
                 node.previousData.reasonCode = node.reason.code;
                 node.previousData.reasonState = node.reason.state;
                 node.previousData.reasonDescription = node.reason.description;
@@ -550,7 +582,7 @@ module.exports = function (RED) {
                     send([null, { topic, payload: timeCtrl }]);
                 }
                 node.previousData.usedRule = ruleId;
-                node.context().set('cacheData', tempData, node.storeName);
+                node.context().set('cacheData', tempData, node.contextStore);
                 if (node.autoTrigger) {
                     node.debug('next autoTrigger will set to ' + node.autoTrigger.time + ' - ' + node.autoTrigger.type);
                     if (node.autoTriggerObj) {
@@ -603,4 +635,24 @@ module.exports = function (RED) {
     }
 
     RED.nodes.registerType('clock-timer', clockTimerNode);
+
+    RED.httpAdmin.post('/clock-timer/:id', RED.auth.needsPermission('clock-timer.write'), (req, res) => {
+        const node = RED.nodes.getNode(req.params.id);
+        if (node !== null && typeof node !== 'undefined') {
+            try {
+                if (req.body && req.body.ruleData) {
+                    changeRules(node, req.body.rulePos, req.body.ruleName, req.body.ruleData);
+                }
+                res.sendStatus(200);
+            } catch(err) {
+                res.sendStatus(500);
+                node.error(RED._('node-red:inject.failed',{error:err.toString()}));
+            }
+        } else {
+            res.status(404).send(JSON.stringify({
+                error: 'can not find node "' +req.params.id+'" '+String(node)
+            }));
+            return;
+        }
+    });
 };